+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # IntelliJ CUE Language Support Changelog
 
 ## [Unreleased]
+- Added: Automatically insert closing quotes when typing the opening quote, available for simple and multiline literals.
 - Fix [#27](https://github.com/nexantic/intellij-cue/issues/27), file starting with attribute raised as invalid
 
 ## [0.5.0]